refactor: move precision and scale logic out of `proof-of-sql-parser` crate
precision and scale are unused within the `proof-of-sql-parser` crate, and as a result, it doesn't make sense to have that logic in the crate.
This commit moves it into the `proof-of-sql` crate via a `IntermediateDecimalExt` trait.

NOTE: this trait will be renamed to `BigDecimalExt` in a future commit.
