+++ crates/proof-of-sql/src/base/math/decimal.rs
@@ -1,5 +1,8 @@
 //! Module for parsing an `IntermediateDecimal` into a `Decimal75`.
-use crate::base::scalar::{Scalar, ScalarConversionError, ScalarExt};
+use crate::base::{
+    math::IntermediateDecimalExt,
+    scalar::{Scalar, ScalarConversionError, ScalarExt},
+};
 use alloc::string::{String, ToString};
 use proof_of_sql_parser::intermediate_decimal::{IntermediateDecimal, IntermediateDecimalError};
 use serde::{Deserialize, Deserializer, Serialize};
