+++ crates/proof-of-sql/src/base/math/mod.rs
@@ -7,3 +7,6 @@ mod log;
 pub(crate) use log::log2_up;
 /// TODO: add docs
 pub(crate) mod permutation;
+
+mod intermediate_decimal_ext;
+pub(crate) use intermediate_decimal_ext::IntermediateDecimalExt;
